+++ Chubberino.UnitTests/Tests/Modules/CheeseGame/PlayerExtensions/WhenAddingPoints.cs
@@ -9,13 +9,11 @@ namespace Chubberino.UnitTests.Tests.Modules.CheeseGame.PlayerExtensions
     public sealed class WhenAddingPoints
     {
         private Player Player { get; }
-        private ICalculator Calculator { get; }
         // private CheeseType Cheese { get; }
 
         public WhenAddingPoints()
         {
             Player = new Player();
-            Calculator = new Calculator();
         }
 
         [Theory]
@@ -33,7 +31,7 @@ namespace Chubberino.UnitTests.Tests.Modules.CheeseGame.PlayerExtensions
 
             Player.Points = initialPoints;
 
-            Player.AddPoints(cheese, Calculator);
+            Player.AddPoints(cheese);
 
             Assert.Equal(expectedPoints, Player.Points);
         }
@@ -53,7 +51,7 @@ namespace Chubberino.UnitTests.Tests.Modules.CheeseGame.PlayerExtensions
             Player.WorkerCount = 1;
             Player.Points = initialPoints;
 
-            Player.AddPoints(cheese, Calculator);
+            Player.AddPoints(cheese);
 
             Assert.Equal(expectedPoints, Player.Points);
         }
@@ -72,7 +70,7 @@ namespace Chubberino.UnitTests.Tests.Modules.CheeseGame.PlayerExtensions
             Player.WorkerCount = workers;
             Player.Points = initialPoints;
 
-            Player.AddPoints(cheese, Calculator);
+            Player.AddPoints(cheese);
 
             Assert.Equal(expectedPoints, Player.Points);
         }
@@ -91,7 +89,7 @@ namespace Chubberino.UnitTests.Tests.Modules.CheeseGame.PlayerExtensions
             Player.WorkerCount = workers;
             Player.Points = initialPoints;
 
-            Player.AddPoints(cheese, Calculator);
+            Player.AddPoints(cheese);
 
             Assert.Equal(0, Player.Points);
         }
@@ -111,7 +109,7 @@ namespace Chubberino.UnitTests.Tests.Modules.CheeseGame.PlayerExtensions
             Player.WorkerCount = workers;
             Player.Points = initialPoints;
 
-            Player.AddPoints(cheese, Calculator);
+            Player.AddPoints(cheese);
 
             Assert.Equal(expectedPoints, Player.Points);
         }
